Support go-to-definition within same file without package declaration
On a file like

	message X {
	}
	message Y {
		X x = 1;
	}

we fail to jump to the definition of X.
This is because we don't search the current file if it has no package.
This seems wrong and unnecessary. The correct criteria is: search
the current file if its package name matches the package identifier
at the cursor. In this case both are empty, which is fine.